+++ backend/characters/signals.go
@@ -7,7 +7,7 @@ import (
 
 var CharacterCreatedSignal = signals.New[*Character]()
 var CharacterUpdatedSignal = signals.New[*Character]()
-var CharacterDeletedSignal = signals.New[int64]()
+var CharacterDeletedSignal = signals.New[int]()
 
 func init() {
 	sse.RegisterSseSourceSignal("CharacterCreated", CharacterCreatedSignal)
